feature:update thread stages in thread details page and fix db query. (#63)
