+++ frontend/src/db/constants.ts
@@ -1,10 +1,16 @@
+// Top level thread status.
+export const STATUS_TODO = "todo";
+export const STATUS_DONE = "done";
+
+// in TODO workflow stages.
 export const NEEDS_NEXT_RESPONSE = "needs_next_response";
 export const NEEDS_FIRST_RESPONSE = "needs_first_response";
 export const HOLD = "hold";
-export const RESOLVED = "resolved";
-export const SPAM = "spam";
 export const WAITING_ON_CUSTOMER = "waiting_on_customer";
-export const TODO = "todo";
 
-export const STATUS_TODO = "todo";
-export const STATUS_DONE = "done";
+// in DONE workflow stages.
+export const RESOLVED = "resolved";
+
+// in other workflow, reserved for later usage.
+// for other status+stages, haven't decided yet.
+export const SPAM = "spam";
